move product service in a separate folder
