+++ lib/product-service/dynamo-tables-construct.ts
