+++ lib/product-service/lambdas/create-product-lambda.ts
