+++ lib/product-service/lambdas/get-product-by-id-lambda.ts
