+++ lib/product-service/lambdas/get-products-lambda.ts
