+++ lib/product-service/lambdas/package.json
