+++ lib/product-service/lambdas/seed-products-lambda.ts
