+++ lib/product-service/product-service-stack.ts
