+++ lib/product-service/product-service.ts
